rosplan_knowledge_base: add constants as instances
Constants were so far ignored by ROSPlan. This in particular means that
they would not be picked up by rqt tools.

This commit adds constants as instances to the knowledge base. While
this fixes most problems, it also means that these instances could be
virtually removed, but still appear in the planning run again. So, this
is to be seen as a quick fix and workaround, not as a full solution
which would require to improve query facilities to also return constants
and mark them as such.